--- vis/v_inputlist
@@ -1,92 +1,0 @@
-/* ************************************************************************** */
-/*                                                                            */
-/*                                                        :::      ::::::::   */
-/*   inputlist.c                                        :+:      :+:    :+:   */
-/*                                                    +:+ +:+         +:+     */
-/*   By: ben <ben@student.42.fr>                    +#+  +:+       +#+        */
-/*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2020/01/14 09:36:33 by bwebb             #+#    #+#             */
-/*   Updated: 2020/05/19 16:29:11 by ben              ###   ########.fr       */
-/*                                                                            */
-/* ************************************************************************** */
-
-#include "../includes/visualiser.h"
-
-int		v_addinputnode(t_input **inputList, char *line)
-{
-	t_input *curnode;
-	t_input *newnode;
-
-	curnode = *inputList;
-	while (curnode && curnode->next)
-		curnode = curnode->next;
-	if (!(newnode = malloc(sizeof(t_input))))
-		return (0);
-	newnode->line = line;
-	newnode->roomnode = NULL;
-	newnode->next = NULL;
-	if (*inputList)
-		curnode->next = newnode;
-	else
-		*inputList = newnode;
-	return (1);
-}
-
-void	v_freeinputlist(t_input *inputlist)
-{
-	if (inputlist)
-	{
-		if (inputlist->next)
-			v_freeinputlist(inputlist->next);
-		if (inputlist->line)
-			free(inputlist->line);
-		if (inputlist->roomnode)
-			freeroom(inputlist->roomnode);
-		free(inputlist);
-	}
-}
-
-void	v_initchecks(t_inputchecks *checks)
-{
-	checks->ants = 0;
-	checks->end = NULL;
-	checks->links = 0;
-	checks->rooms = 0;
-	checks->start = NULL;
-}
-
-int		v_runinputchecks(t_heart *heart)
-{
-	t_input *inputnode;
-	t_inputchecks *inputchecks;
-
-	inputnode = heart->input;
-	inputchecks =  heart->inputchecks;
-	while (inputnode)
-	{
-
-		if (ft_strequ(inputnode->line, "##start") == 1)
-		{
-			if ((!inputnode->next) || (!isroom(inputnode->next->line, heart)) || inputchecks->start)
-				return (0);
-			inputchecks->start = &inputnode->next->roomnode;
-		}
-		else if (ft_strequ(inputnode->line, "##end") == 1)
-		{
-			if ((!inputnode->next) || (!isroom(inputnode->next->line, heart)) || inputchecks->end)
-				return (0);
-			inputchecks->end = &inputnode->next->roomnode;
-		}
-		inputnode = inputnode->next;
-	}
-	return (1);
-}
-
-int		v_validateinput(t_heart *heart)
-{
-	if (!(heart->inputchecks = malloc(sizeof(t_inputchecks))))
-		erexit(heart, 2);
-	v_initchecks(heart->inputchecks);
-	return (v_runinputchecks(heart));
-	//dupe links
-}
